+++ flag/guest/guest.go
@@ -1,6 +1,6 @@
 package guest
 
-import "github.com/giantswarm/cluster-operator/v4/flag/guest/cluster"
+import "github.com/giantswarm/cluster-operator/v5/flag/guest/cluster"
 
 // Guest is a data structure to hold guest cluster specific configuration
 // flags.